# evidence/ex13_capstone/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "API de Agendamento Clínico",
    "description": "Pacientes, profissionais de saúde e consultas. Dados de saúde (LGPD art. 11).",
    "version": "1.0.0"
  },
  "paths": {
...
    "/appointment/new": {
      "post": {
        "tags": [
          "Appointments"
        ],
        "summary": "Create Appointment",
        "description": "Somente profissionais agendam, e só para os próprios pacientes. O\n`professional_id` é o do token — o cliente não pode escolher outro.",
        "operationId": "create_appointment_appointment_new_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AppointmentCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppointmentPublic"
                }
              }
            }
          },
          "401": {
            "description": "Token ausente, inválido ou expirado"
          },
          "403": {
            "description": "Perfil/escopo sem permissão para a operação"
          },
          "429": {
            "description": "Limite de requisições excedido (veja Retry-After)"
          },
          "404": {
            "description": "Recurso inexistente OU pertencente a outro usuário (indistinguíveis de propósito)"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/appointment/": {
      "get": {
        "tags": [
          "Appointments"
        ],
        "summary": "Read Appointments",
        "operationId": "read_appointments_appointment__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/AppointmentStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "date_from",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date From"
            }
          },
          {
            "name": "date_to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date To"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AppointmentPublic"
                  },
                  "title": "Response Read Appointments Appointment  Get"
                }
              }
            }
          },
          "401": {
            "description": "Token ausente, inválido ou expirado"
          },
          "403": {
            "description": "Perfil/escopo sem permissão para a operação"
          },
          "429": {
            "description": "Limite de requisições excedido (veja Retry-After)"
          },
          "404": {
            "description": "Recurso inexistente OU pertencente a outro usuário (indistinguíveis de propósito)"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/appointment/agenda": {
      "get": {
        "tags": [
          "Appointments"
        ],
        "summary": "Read Agenda",
        "description": "Agenda do dia (JSON). Sem anotações clínicas: serve também à recepção.",
        "operationId": "read_agenda_appointment_agenda_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "day",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Day"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AgendaItem"
                  },
                  "title": "Response Read Agenda Appointment Agenda Get"
                }
              }
            }
          },
          "401": {
            "description": "Token ausente, inválido ou expirado"
          },
          "403": {
            "description": "Perfil/escopo sem permissão para a operação"
          },
          "429": {
            "description": "Limite de requisições excedido (veja Retry-After)"
          },
          "404": {
            "description": "Recurso inexistente OU pertencente a outro usuário (indistinguíveis de propósito)"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/appointment/{appointment_id}": {
      "get": {
        "tags": [
          "Appointments"
        ],
        "summary": "Read Appointment",
        "operationId": "read_appointment_appointment__appointment_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "appointment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "exclusiveMinimum": 0,
              "title": "Appointment Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppointmentPublic"
                }
              }
            }
          },
          "401": {
            "description": "Token ausente, inválido ou expirado"
          },
          "403": {
            "description": "Perfil/escopo sem permissão para a operação"
          },
          "429": {
            "description": "Limite de requisições excedido (veja Retry-After)"
          },
          "404": {
            "description": "Recurso inexistente OU pertencente a outro usuário (indistinguíveis de propósito)"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Appointments"
        ],
        "summary": "Update Appointment",
        "operationId": "update_appointment_appointment__appointment_id__put",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "appointment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "exclusiveMinimum": 0,
              "title": "Appointment Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AppointmentUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppointmentPublic"
                }
              }
            }
          },
          "401": {
            "description": "Token ausente, inválido ou expirado"
          },
          "403": {
            "description": "Perfil/escopo sem permissão para a operação"
          },
          "429": {
            "description": "Limite de requisições excedido (veja Retry-After)"
          },
          "404": {
            "description": "Recurso inexistente OU pertencente a outro usuário (indistinguíveis de propósito)"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Appointments"
        ],
        "summary": "Cancel Appointment",
        "description": "Cancelamento lógico (mantém o histórico para auditoria).",
        "operationId": "cancel_appointment_appointment__appointment_id__delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "appointment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "exclusiveMinimum": 0,
              "title": "Appointment Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppointmentPublic"
                }
              }
            }
          },
          "401": {
            "description": "Token ausente, inválido ou expirado"
          },
          "403": {
            "description": "Perfil/escopo sem permissão para a operação"
          },
          "429": {
            "description": "Limite de requisições excedido (veja Retry-After)"
          },
          "404": {
            "description": "Recurso inexistente OU pertencente a outro usuário (indistinguíveis de propósito)"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AgendaItem": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "date_time": {
            "type": "string",
            "format": "date-time",
            "title": "Date Time"
          },
          "status": {
            "$ref": "#/components/schemas/AppointmentStatus"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          },
          "patient_name": {
            "type": "string",
            "title": "Patient Name"
          },
          "professional_name": {
            "type": "string",
            "title": "Professional Name"
          }
        },
        "type": "object",
        "required": [
          "id",
          "date_time",
          "status",
          "reason",
          "patient_name",
          "professional_name"
        ],
        "title": "AgendaItem",
        "description": "Visão da recepção: sem anotações clínicas, sem CPF, sem telefone."
      },
      "AppointmentCreate": {
        "properties": {
          "patient_id": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Patient Id"
          },
          "date_time": {
            "type": "string",
            "format": "date-time",
            "title": "Date Time"
          },
          "reason": {
            "type": "string",
            "pattern": "^[A-Za-z0-9À-ÖØ-öø-ÿ\\s.,;:!?()\\-_/%+ºª°'\\\"]{3,120}$",
            "title": "Reason"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^[A-Za-z0-9À-ÖØ-öø-ÿ\\s.,;:!?()\\-_/%+ºª°'\\\"]{0,2000}$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "patient_id",
          "date_time",
          "reason"
        ],
        "title": "AppointmentCreate"
      },
      "AppointmentPublic": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "patient_id": {
            "type": "integer",
            "title": "Patient Id"
          },
          "professional_id": {
            "type": "integer",
            "title": "Professional Id"
          },
          "date_time": {
            "type": "string",
            "format": "date-time",
            "title": "Date Time"
          },
          "status": {
            "$ref": "#/components/schemas/AppointmentStatus"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          }
        },
        "type": "object",
        "required": [
          "id",
          "patient_id",
          "professional_id",
          "date_time",
          "status",
          "reason"
        ],
        "title": "AppointmentPublic",
        "description": "Whitelist de campos devolvidos pela API. created_by_user_id, created_at,\nupdated_at, created_from_ip e internal_audit_note ficam de fora de propósito."
      },
      "AppointmentStatus": {
        "type": "string",
        "enum": [
          "agendada",
          "cancelada",
          "concluida"
        ],
        "title": "AppointmentStatus"
      },
      "AppointmentUpdate": {
        "properties": {
          "date_time": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Date Time"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^[A-Za-z0-9À-ÖØ-öø-ÿ\\s.,;:!?()\\-_/%+ºª°'\\\"]{3,120}$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^[A-Za-z0-9À-ÖØ-öø-ÿ\\s.,;:!?()\\-_/%+ºª°'\\\"]{0,2000}$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AppointmentStatus"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "AppointmentUpdate"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/user/signin"
          }
        }
      },
      "LabClientCredentials": {
        "type": "oauth2",
        "flows": {
          "clientCredentials": {
            "scopes": {
              "availability:read": "Consultar horários livres de um profissional"
            },
            "tokenUrl": "/oauth/token"
          }
        }
      }
    }
  }
}